GenAi/Embedder/requirements.txt
accelerate
azure-identity
azure-search-documents==11.4.0b8
boto3
chromadb
langchain==0.0.335
opensearch-py
pgvector
psycopg2
pyinstaller
python-dotenv
pypdf
redis
requests_aws4auth
sentence_transformers
tiktoken
torch
transformers
openai==0.27.8
azure-ai-formrecognizer
transformers 
torch
flask
flask-cors